Fix handling of x18 in AArch64 inline assembly on ohos/trusty or with -Zfixed-x18
